+++ src/main/java/pascal/taie/analysis/pta/plugin/taint/FieldSource.java
@@ -25,7 +25,8 @@ package pascal.taie.analysis.pta.plugin.taint;
 import pascal.taie.language.classes.JField;
 import pascal.taie.language.type.Type;
 
-public record FieldSource(JField field, Type type) implements Source {
+public record FieldSource(JField field, Type type, String rawEntry)
+        implements Source {
 
     @Override
     public String toString() {